blog_platform/requirements.txt
asgiref==3.7.2
awscli==1.32.76
boto3==1.34.76
botocore==1.34.76
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.4
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.11.2
django-health-check==3.18.2
django-storages==1.14.2
djangorestframework==3.14.0
docutils==0.16
gunicorn==22.0.0
idna==3.6
jmespath==1.0.1
packaging==24.0
Pillow==10.1.0
psycopg2==2.8.6
psycopg2-binary==2.9.9
pyasn1==0.6.0
python-dateutil==2.9.0.post0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rsa==4.7.2
s3transfer==0.10.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
urllib3>=1.25.4,<1.27
